\name{CLOU.CLAY.SALL}
\alias{CLOU.CLAY.SALL}

\title{
	Weekly maximum wind speed data collected over the CLOU, CLAY and SALL stations located in Oklahoma, USA, over the March-May preiod between 1996 and 2012.
}

\description{
	Weekly maximum wind speed data at CLOU, CLAY and SALL stations. 
}

\details{
	it is a \code{data.frame} object with \eqn{3}{3} columns and \eqn{212}{212} rows.	
}

\references{

	Beranger, B., Padoan, S. A. and Sisson, S. A. (2016).
	Models for extremal dependence derived from skew-symmetric families.
	\emph{Scandinavian Journal of Statistics}, \bold{To appear}.	
	
}
