\name{ExtremalDep-internal}

\alias{matrix.trace}
\alias{matrix.sqrt}
\alias{FitGev}
\alias{Dist2Dist}
\alias{GevLogLik}
\alias{MomEst}

\title{Internal functions in the ExtremalDep library}
\description{
These functions are user-level but which the user is not required to use directly.
}


\keyword{internal}
