\name{logReturns}
\alias{logReturns}

\title{Monthly maxima of log-return exchange rates of the Pound Sterling (GBP)
against the US dollar (USD) and the Japanese yen (JPY), between March 1991 and
December 2014}

\description{
  The dataset \code{logReturns} contains four columns:
  \code{date_USD} and \code{USD} give the date and value of the monthly maxima
  of the log-return exchange rate GBP/USD, while \code{date_JPY} and \code{JPY}
  give the date and value of the monthly maxima of the log-return exchange rate
  GBP/JPY.
}

\format{
  A \eqn{286 \times 4} matrix. The first and third columns are of type
  \code{"character"}, while the second and fourth columns are of type
  \code{"numeric"}.
}

\keyword{datasets}
