\name{MilanPollution}
\alias{Milan.winter}
\alias{Milan.summer}

\title{Pollution data for summer and winter months in Milan, Italy.}

\description{Two datasets \code{Milan.summer} and \code{Milan.winter}, each containing 5 air pollutants: daily maximum of NO2, NO, O3 and SO2, daily mean of PM10; and 6 meteorological covariates: maximum precipitation, maximum temperature, maximum humidity, mean precipitation, mean temperature and mean humidity.}

\details{The summer period corresponds to the period 30 April - 30 August between 2003 and 2017 and thus the dataset contains \eqn{1968}{1968} observations. The winter period corresponds to the period 32 November - 27(28) February. The records start from 31 December 2001 until 30 December 2017 and thus the dataset contains \eqn{1924}{124} observations.}

\format{A \eqn{1968*12}{1968*12} data frame and a \eqn{1924*12}{1924*12} data frame.}

\keyword{datasets}
