\name{winterdat}
\alias{winterdat}
\alias{print.winterdat}

\title{\eqn{5}{5}-dimensional air quality dataset recorded in Leeds (U.K.), during five winter seasons.}
\description{The dataset contains \eqn{590}{590} (transformed) daily maxima of five 
	air pollutants (respectively PM10, NO, NO2, 03 and S02) recorded in Leeds (U.K.) during five winter 
	seasons (November-Februrary) between 1994 and 1998. Contains \code{NA}'s. 
	Outliers have been removed according to Heffernan and Tawn (2004).}
\format{A \eqn{590*5}{590*5} matrix.}
\source{\url{http://airquality.co.uk}}
\references{
	
	Heffernan, J. E., and Tawn, J. A. (2004).
	A conditional approach for multivariate extreme values.
	\emph{Journal of the Royal Statistical Society, Series B, Methodology}, \bold{66}, 497--546
	
}
