\name{logReturns}
\alias{logReturns}

\title{Monthly maxima of log-return exchange rates of the Pound Sterling (GBP) against the US dollar (USD) and the Japanese yen (JPY), between March 1991 and December 2014.}

\description{The dataset \code{logReturns} contains 4 columns: \code{date_USD} and \code{USD} give the date of the monthly maxima of the log-return exchange rate GBP/USD and its value while \code{date_JPY} and \code{JPY} give the date of the monthly maxima of the log-return exchange rate GBP/JPY and its value.}

\format{A \eqn{286*4}{286*4} matrix. The first and third columns are objects of type \code{"character"} while the second and fourth columns are of type \code{"numeric"}.}

\keyword{datasets}
