\name{PrecipFrance}
\alias{PrecipFrance}
\alias{precip}
\alias{lat}
\alias{lon}

\title{Weekly maxima of hourly rainfall in France}

\description{List containing the weekly maxima of hourly rainfall in the Fall season from 1993 to 2011 recorded at 92 stations across France (\code{precip}). Coordinates of the monitoring stations are given in \code{lat} and \code{lon}. 
}

\details{The fall season corresponds to the September-November (SON) period.  The data thus cover a 12-week period over \eqn{19}{19} years, yielding a sample of \eqn{n=228}{n=228} observations (rows) and \eqn{p=92}{p=92} stations (columns).}

\format{A list made of a \eqn{228*92}{228*92} matrix (\code{precip}) and two vectors of length \eqn{228}{228} (\code{lat} and \code{lon}).}

\keyword{datasets}
