\name{Wind}
\alias{Wind}

\title{Weekly maximum wind speed data collected over 4 stations across Oklahoma, USA, over the March-May preiod between 1996 and 2012.}
\description{
	There are four datasets of weekly maximum wind speed data, for each triple of location: \code{CLOU.CLAY.SALL}, \code{CLOU.CLAY.PAUL}, \code{CLAY.SALL.PAUL} and \code{CLOU.SALL.PAUL}. 
}


\details{
	\code{CLOU.CLAY.SALL} is a \code{data.frame} object	with \eqn{3}{3} columns and \eqn{212}{212} rows.	
	\code{CLOU.CLAY.PAUL} is a \code{data.frame} object	with \eqn{3}{3} columns and \eqn{217}{217} rows.	
	\code{CLAY.SALL.PAUL} is a \code{data.frame} object	with \eqn{3}{3} columns and \eqn{211}{211} rows.	
	\code{CLOU.SALL.PAUL} is a \code{data.frame} object	with \eqn{3}{3} columns and \eqn{217}{217} rows.				
	Missing observations have been discarded for each triple.
}

\references{

	Beranger, B., Padoan, S. A. and Sisson, S. A. (2016).
	Models for extremal dependence derived from skew-symmetric families.
	\emph{Scandinavian Journal of Statistics}, \bold{To appear}.	
	
}
