\name{Wind}
\alias{Wind}
\alias{CLAY.SALL.PAUL}
\alias{CLOU.CLAY.PAUL}
\alias{CLOU.CLAY.SALL}
\alias{CLOU.SALL.PAUL}

\title{Weekly maximum wind speed data collected over 4 stations across Oklahoma, USA, over the March-May preiod between 1996 and 2012.}
\description{
	There are four datasets of weekly maximum wind speed data, for each triplet of locations: \code{CLOU.CLAY.SALL}, \code{CLOU.CLAY.PAUL}, \code{CLAY.SALL.PAUL} and \code{CLOU.SALL.PAUL}.
}


\details{
	\code{CLOU.CLAY.SALL} is a \code{data.frame} object	with \eqn{3}{3} columns and \eqn{212}{212} rows.
	\code{CLOU.CLAY.PAUL} is a \code{data.frame} object	with \eqn{3}{3} columns and \eqn{217}{217} rows.
	\code{CLAY.SALL.PAUL} is a \code{data.frame} object	with \eqn{3}{3} columns and \eqn{211}{211} rows.
	\code{CLOU.SALL.PAUL} is a \code{data.frame} object	with \eqn{3}{3} columns and \eqn{217}{217} rows.
	Missing observations have been discarded for each triplet.
}

\references{

	Beranger, B., Padoan, S. A. and Sisson, S. A. (2017).
	Models for extremal dependence derived from skew-symmetric families.
	\emph{Scandinavian Journal of Statistics}, \bold{44}(1), 21-45.

}

\keyword{datasets}
